Fixed a serialization/deserialization bug: the recruitment object is now serialized/deserialization when the simulation state is saved and loaded. Otherwise the state of a lymph ode recruitment object won't be restored when a state is loaded. Updated the gui version accordingly.
git-svn-id: svn://innoculant.micro.med.umich.edu/dev/gr2d/GR-ABM-ODE@1457 88ce85bb-8564-4ae6-884f-ab7503230054

+++ vectordatasets/vectorgrid.cpp
@@ -13,9 +13,9 @@ VectorGrid::VectorGrid(size_t _DIM)
 	: Grid(_DIM)
 	, _grid(_DIM*_DIM)
 {
-	for (int i = 0; i < _DIM; i++) // row
+	for (size_t i = 0; i < _DIM; i++) // row
 	{
-		for (int j = 0; j < _DIM; j++) // col
+		for (size_t j = 0; j < _DIM; j++) // col
 		{
 			VectorGridItem& item = _grid[j + i * _DIM];
